+++ packages/react/components/list-input.js
@@ -377,6 +377,18 @@ class F7ListInput extends React.Component {
   }
 
   componentDidUpdate(prevProps, prevState) {
+    __reactComponentWatch(this, 'props.colorPickerParams', prevProps, prevState, () => {
+      const self = this;
+      if (!self.$f7 || !self.f7ColorPicker) return;
+      Utils.extend(self.f7ColorPicker.params, self.colorPickerParams || {});
+    });
+
+    __reactComponentWatch(this, 'props.calendarParams', prevProps, prevState, () => {
+      const self = this;
+      if (!self.$f7 || !self.f7Calendar) return;
+      Utils.extend(self.f7Calendar.params, self.f7Calendar || {});
+    });
+
     __reactComponentWatch(this, 'props.value', prevProps, prevState, () => {
       const self = this;
       if (!self.$f7) return;